% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_curl_options}
\alias{add_curl_options}
\title{Add Options}
\usage{
add_curl_options(req, curl_options)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{options}{List of arguments to `httr2::req_options`}
}
\description{
Add list of arguments for `httr2::req_options` to a request object
}
