% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-wrappers.R
\name{add_error_handling}
\alias{add_error_handling}
\title{Add Error Handling}
\usage{
add_error_handling(req, error_handling_options = NULL)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{headers}{List of arguments to \code{\link[httr2:req_error]{httr2::req_error()}}}
}
\description{
Add list of arguments to a request object that when errors occur and how the message is
extracted.
}
