% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_headers}
\alias{add_headers}
\title{Add Headers}
\usage{
add_headers(req, headers)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{headers}{List of arguments to `httr2::req_headers`}
}
\description{
Add list of arguments for `httr2::req_headers` to a request object
}
