% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_proxy}
\alias{add_proxy}
\title{Add Proxy Args}
\usage{
add_proxy(req, proxy_args)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{proxy_args}{List of arguments to `httr2::req_proxy`}
}
\description{
Add list of arguments for `httr2::req_proxy` to a request object
}
