% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_retry}
\alias{add_retry}
\title{Add Retry}
\usage{
add_retry(req, options)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{headers}{List of arguments to `httr2::req_retry`}
}
\description{
Add list of arguments for `httr2::req_retry` to a request object
}
