% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2-wrappers.R
\name{add_throttle}
\alias{add_throttle}
\title{Add Throttle}
\usage{
add_throttle(req, throttle_rate)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{throttle_rate}{List of arguments to \link[httr2:req_throttle]{httr2::req_throttle}}
}
\description{
Add list of throttle arguments to a request object
}
\keyword{internal}
