% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2-wrappers.R
\name{add_to_path}
\alias{add_to_path}
\title{Add To URL Path}
\usage{
add_to_path(req, paths)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{paths}{List of arguments to \code{\link[httr2:req_url]{httr2::req_url_path_append()}}}
}
\description{
Add To URL Path
}
\keyword{internal}
