% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_df}
\alias{check_df}
\title{Check the classes of a data frame and allow for multiple different classes for certain columns}
\usage{
check_df(df, class_list)
}
\arguments{
\item{df}{The data frame to check}

\item{class_list}{A named list that specifies the expected classes for each column

The \code{class_list} argument should be a named list, where the names correspond to the column
names in \code{df}. Each entry in the list should be a character vector of the allowed classes for
the corresponding column. For example, \code{class_list = list(col1 = "numeric", col2 = c("character", "factor"), col3 = "logical")} specifies that \code{col1} must be of class
\code{"numeric"}, \code{col2} must be of class \code{"character"} or \code{"factor"}, and \code{col3} must be of class
\code{"logical"}.}
}
\value{
The checked data frame
}
\description{
This function checks the classes of a data frame and allows for certain columns to have multiple
allowed classes.
}
\examples{
\dontrun{
df <- data.frame(
  col1 = c(1, 2, 3),
  col2 = c("a", "b", "c"),
  col3 = c(TRUE, FALSE, TRUE),
  stringsAsFactors = FALSE
)

class_list <- list(
  col1 = "numeric",
  col2 = c("character", "factor"),
  col3 = "logical"
)

check_df(df, class_list)
}

}
\keyword{internal}
