% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftgroup.R
\name{check_draft_group_id}
\alias{check_draft_group_id}
\title{Check Draft Group Inputs}
\usage{
check_draft_group_id(draft_group_id, contest_id)
}
\description{
Use either \code{draft_group_id} or \code{contest_id} if \code{draft_group_id} is \code{NULL}.
}
