% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_solver}
\alias{check_solver}
\title{Check Solver}
\usage{
check_solver(solver)
}
\arguments{
\item{solver}{A string of the solver name.}
}
\description{
Ensure the solver name passed to \code{dk_optimize_lineup} is available
}
\keyword{internal}
