% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_multi_get.R
\name{chunk}
\alias{chunk}
\title{Chunk a vector into a specified number of chunks or a specified chunk size}
\usage{
chunk(x, n_chunks = NULL, chunk_size = NULL)
}
\arguments{
\item{x}{A vector or other R object that can be split.}

\item{n_chunks}{An optional integer indicating the desired number of chunks.}

\item{chunk_size}{An optional integer indicating the desired number of elements per chunk.}
}
\value{
A list of chunks.
}
\description{
Chunk a vector into a specified number of chunks or a specified chunk size
}
\details{
This function splits the input \code{x} into multiple chunks based on the specified
\code{n_chunks} or \code{chunk_size}. If both are provided, \code{chunk_size} takes precedence
and \code{n_chunks} is ignored with a warning. If neither is provided, the function
will stop with an error. The function returns a list of chunks.
}
\examples{
vec <- 1:10
chunk(vec, chunk_size = 3)
chunk(vec, n_chunks = 3)

}
\keyword{internal}
