% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{dk_extract_solution}
\alias{dk_extract_solution}
\title{Extract Solution}
\usage{
dk_extract_solution(solved_model)
}
\arguments{
\item{solved_model}{Optimal lineup object created by \code{\link[=dk_optimize_lineup]{dk_optimize_lineup()}}.}
}
\description{
Extract information from optimized lineup solution
}
