% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{dk_get_competitions}
\alias{dk_get_competitions}
\title{Get Competitions for a Draft Group}
\usage{
dk_get_competitions(
  draft_group_id = NULL,
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{draft_group_id}{Sequence of digits that correspond to a draft table/group.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{
  Arguments passed on to \code{\link[=dk_get_contest_info]{dk_get_contest_info}}
  \describe{
    \item{\code{contest_key}}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}
  }}
}
\description{
Fetch information on the competitions associated with a specific
draft group ID. In particular, this includes home/away team designation,
team IDs, and sport name.
}
\examples{
  \dontrun{
    # NFL Showdown Captain Mode
    dk_get_competitions(75284)

    # NFL Classic
    dk_get_competitions(74625)

    # Multiple competitions
    dk_get_competition(42994)
  }

}
