% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftgroup.R
\name{dk_get_draft_group}
\alias{dk_get_draft_group}
\title{Get Draft group}
\usage{
dk_get_draft_group(
  draft_group_id = NULL,
  contest_key = NULL,
  draftable_ids = NULL,
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{draft_group_id}{Sequence of digits that correspond to a draft table/group.
If \code{draft_group_id} and \code{contest_key} are both passed, \code{contest_key} is ignored.}

\item{contest_key}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{draftable_ids}{Unique draftable IDs to filter by. If \code{NULL} (the default),
then all draftable IDs will be included in the result.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{
  Arguments passed on to \code{\link[=dk_get_contest_info]{dk_get_contest_info}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Fetch information on the available players for a specific
draft group ID. In particular, this includes player names/IDs and
their salaries.
}
\examples{
  \dontrun{
    get_draft_group(75367)
  }

}
