% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaderboard.R
\name{dk_get_entries}
\alias{dk_get_entries}
\title{Get Entry Details}
\usage{
dk_get_entries(
  draft_group_id,
  entry_keys,
  cookies = path.expand("~/cookies.json"),
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{draft_group_id}{Sequence of digits that correspond to a draft table/group.
If \code{draft_group_id} and \code{contest_key} are both passed, \code{contest_key} is ignored.}

\item{entry_keys}{Vector of numeric (or character) keys that correspond to a specific entry in
a specific contest. See output from \code{\link[=dk_get_leaderboard]{dk_get_leaderboard()}}.}

\item{cookies}{Path to JSON of cookies needed to perform API request.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{
  Arguments passed on to \code{\link[=dk_request]{dk_request}}
  \describe{
    \item{\code{query_params}}{A list of query parameters passed to \verb{[httr2::req_url_query]}.}
    \item{\code{proxy_args}}{List of arguments to \link[httr2:req_proxy]{httr2::req_proxy}}
    \item{\code{curl_options}}{List of arguments to \link[httr2:req_options]{httr2::req_options}}
    \item{\code{throttle_rate}}{List of arguments to \link[httr2:req_throttle]{httr2::req_throttle}}
    \item{\code{headers}}{List of arguments to \code{\link[httr2:req_headers]{httr2::req_headers()}}}
    \item{\code{paths}}{List of arguments to \code{\link[httr2:req_url]{httr2::req_url_path_append()}}}
    \item{\code{error_handling_options}}{List of arguments to \code{\link[httr2:req_error]{httr2::req_error()}}}
    \item{\code{retry_options}}{List of arguments passed to \code{\link[httr2:req_retry]{httr2::req_retry()}}}
    \item{\code{base_url}}{Base URL for request.}
    \item{\code{method}}{Custom HTTP method}
  }}
}
\description{
Fetch details for entries in a contest,
including the drafted roster for each
entry, stats for each player in an entry roster,
and the fantasy points associated to each stat.
}
\examples{
  \dontrun{
    dk_get_entries(draft_group_id = 80584, entry_keys = c(3618408508, 3618897002))
  }

}
