% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{dk_get_game_type_rules}
\alias{dk_get_game_type_rules}
\title{Get Gametype Rules}
\usage{
dk_get_game_type_rules(
  game_type_id = NULL,
  contest_key = NULL,
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{game_type_id}{Integer corresponding to the game type.
For example, 159 in \url{https://api.draftkings.com/lineups/v1/gametypes/159/rules}.
If both \code{game_type_id} and \code{contest_key} are passed, then \code{contest_key} is ignored.}

\item{contest_key}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{Arguments passed to \code{\link[=dk_request]{dk_request()}}}
}
\description{
Fetch rules corresponding to a specific game type ID.
}
