% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{dk_get_lobby_contests}
\alias{dk_get_lobby_contests}
\title{Get List of Contests in Lobby}
\usage{
dk_get_lobby_contests(
  sport = NULL,
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{sport}{character. optional.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{Arguments passed to \code{\link[=dk_request]{dk_request()}}}
}
\description{
Fetch the full table of contests and related info from DraftKings.com lobby
}
