% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{dk_get_lobby_game_types}
\alias{dk_get_lobby_game_types}
\title{Get List of Game Types in Lobby}
\usage{
dk_get_lobby_game_types(
  sport = NULL,
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{sport}{character. optional.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{
  Arguments passed on to \code{\link[=dk_get_lobby_contests]{dk_get_lobby_contests}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Fetch the full list of game types
}
