% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{dk_get_optimal_lineups}
\alias{dk_get_optimal_lineups}
\title{Get Multiple Optimal Lineups}
\usage{
dk_get_optimal_lineups(schematic, n = 1, tolerance = 0.001)
}
\arguments{
\item{schematic}{Output from \code{\link[=dk_prepare_schematic]{dk_prepare_schematic()}} which
includes information needed for optimization; including player info such as,
news status, salary, and projected fantasy points.}

\item{n}{Number of lineups to return.}

\item{tolerance}{The difference in fantasy points between the previous
optimal lineup and the current optimal lineup during iterations.
Usually the default of 0.001 is fine.}
}
\description{
Wrapper around \code{\link[=dk_optimize_lineup]{dk_optimize_lineup()}}. Can return an
arbitrary number of optimal lineups in succession.
}
