% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftgroup.R
\name{dk_get_player_fp}
\alias{dk_get_player_fp}
\title{Get Player Fantasy Points Earned}
\usage{
dk_get_player_fp(
  timeframe,
  season = as.numeric(format(Sys.Date(), "\%Y")),
  sport = c("nfl", "nba", "mlb"),
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{timeframe}{integer. Either the week number for NFL, or a date of the form \code{20230312} for
MLB and NBA. If the timeframe format detected does not match the sport argument passed, an
error is returned.}

\item{season}{integer. optional. Defaults to the current season year.}

\item{sport}{character. Defaults to NFL.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{
  Arguments passed on to \code{\link[=dk_get_contest_info]{dk_get_contest_info}}
  \describe{
    \item{\code{contest_key}}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}
  }}
}
\description{
Retrieve player fantasy points earned in each
game for a given season and week (NFL) or date (NBA/MLB).
}
\details{
Data starts on October 18th, 2022 for NBA.
}
