% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_multi_get.R
\name{dk_multi_get}
\alias{dk_multi_get}
\title{Get Draft Kings Data in Parallel}
\usage{
dk_multi_get(
  reqs,
  paths = NULL,
  chunk_size = Inf,
  n_chunks = NULL,
  sleep = 3,
  report_time = TRUE,
  .progress = TRUE,
  ...
)
}
\arguments{
\item{reqs}{A list of \link[httr2]{request}s.}

\item{paths}{An optional list of paths, if you want to download the request
bodies to disks. If supplied, must be the same length as \code{reqs}.}

\item{chunk_size}{An optional integer indicating the desired number of elements per chunk.}

\item{n_chunks}{An optional integer indicating the desired number of chunks.}

\item{sleep}{A numeric value indicating the number of seconds to sleep
after performing the requests. If NULL or negative, no sleep is performed.}

\item{report_time}{Should the time it took the function to complete be reported?}

\item{.progress}{Whether to show a progress bar. Use \code{TRUE} to turn on
a basic progress bar, use a string to give it a name, or see
\link[purrr]{progress_bars} for more details.}

\item{...}{Other arguments passed to \code{\link[=dk_req_perform]{dk_req_perform()}}}
}
\description{
Fetch DraftKings data in parallel using \code{httr2::multi_req_perform()}
and in batches.
}
