% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{dk_optimize_lineup.classic}
\alias{dk_optimize_lineup.classic}
\title{Optimize Classic Game Type Lineup}
\usage{
\method{dk_optimize_lineup}{showdown_captain_mode}(schematic, max_points = NULL, ...)
}
\arguments{
\item{schematic}{Output from \verb{[dk_prepare_schematic()]} which
includes information needed for optimization; including player info such as,
news status, salary, and projected fantasy points.}

\item{max_points}{Optional upper threshold of points. This is used primarily
to get many successive lineups in \verb{[get_optimal_lineups()]}}

\item{...}{Other arguments passed to optimization method.}
}
\description{
Determine the optimal lineup of players based on the list of
available players, their salaries, and their projected
fantasy points for a classic game type.
}
