% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{dk_req_process}
\alias{dk_req_process}
\title{Process Draft Kings Request Object}
\usage{
dk_req_process(
  req,
  output = c("cleaned_json", "json", "response", "request"),
  objclass = NULL,
  ...
)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{objclass}{character. Name of class to assign to output object. Default
is NULL, and no additional class is assigned. Used to facilitate parsing;
See \code{\link[=dk_resp_parse]{dk_resp_parse()}}.}

\item{...}{Other arguments passed to \code{\link[=dk_req_perform]{dk_req_perform()}}}
}
\value{
See \code{output}.
}
\description{
Wrapper to facilitate how a request should
be processed based on a desired output.
}
