% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{dk_write_lineups_to_csv}
\alias{dk_write_lineups_to_csv}
\title{Write Lineups to CSV}
\usage{
dk_write_lineups_to_csv(optimal_lineups, file = "lineups.csv")
}
\arguments{
\item{optimal_lineups}{Output from \verb{[get_optimal_lineups()]}}

\item{file}{Path and name of the CSV file to create}
}
\description{
Given the output from \verb{[dk_get_optimal_lineups()]}, create a
CSV formatted for uplaod to \url{https://www.draftkings.com/lineup/upload}
}
