% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkreq.R
\name{dkreq}
\alias{dkreq}
\title{Create Draft Kings Request Object}
\usage{
dkreq(
  proxy_args = NULL,
  curl_options = NULL,
  throttle_rate = NULL,
  headers = NULL,
  retry_options = NULL,
  url_path = NULL,
  query_params = NULL,
  error_handling_options = NULL,
  base_url = "https://api.draftkings.com/"
)
}
\arguments{
\item{proxy_args}{List of arguments to \link[httr2:req_proxy]{httr2::req_proxy}}

\item{throttle_rate}{List of arguments to \link[httr2:req_throttle]{httr2::req_throttle}}

\item{headers}{List of arguments to \code{\link[httr2:req_headers]{httr2::req_headers()}}}

\item{base_url}{Base URL for request.}

\item{is_error_func}{See \code{\link[httr2:req_error]{httr2::req_error()}}}

\item{error_body_func}{See \code{\link[httr2:req_error]{httr2::req_error()}}}
}
\description{
Create Draft Kings Request Object
}
