% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.kings-package.R
\docType{package}
\name{draft.kings-package}
\alias{draft.kings}
\alias{draft.kings-package}
\title{draft.kings: Draft Kings API Wrapper and Optimization}
\description{
Fetch data from DraftKings using the provided API and perform lineup optimization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gacolitti.github.io/draft.kings}
  \item Report bugs at \url{https://github.com/gacolitti/draft.kings/issues}
}

}
\keyword{internal}
