% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{format_duration}
\alias{format_duration}
\title{Format a time duration into a human-readable time format}
\usage{
format_duration(start_time, end_time = NULL)
}
\arguments{
\item{start_time}{A POSIXct value representing the start time of the duration}

\item{end_time}{A POSIXct value representing the end time of the duration (default is
NULL, which sets the end time to the current system time)}
}
\value{
A character string representing the formatted duration
}
\description{
Format a time duration into a human-readable time format
}
\keyword{internal}
