% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{format_time}
\alias{format_time}
\title{Format a given number of seconds into a human-readable time format}
\usage{
format_time(seconds, digits = 1)
}
\arguments{
\item{seconds}{A numeric value representing the number of seconds to format}

\item{digits}{An integer value indicating the number of digits after the decimal point
to round to (default is 1)}
}
\value{
A character string representing the formatted time
}
\description{
Format a given number of seconds into a human-readable time format
}
\keyword{internal}
