% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{get_contest_schematic}
\alias{get_contest_schematic}
\title{Get Contest Schematic}
\usage{
get_contest_schematic(
  contest_id,
  player_df = NULL,
  rules = NULL,
  players = NULL,
  exclude_players = NULL,
  exclude_questionable = TRUE
)
}
\arguments{
\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{player_df}{Object returned by \code{get_player_list()}. If \code{NULL} (the default),
then this object is fetched using the \code{contest_id}.}

\item{rules}{Object returned by \code{get_gametype_rules()}. If \code{NULL} (the default),
then this object is fetched using the \code{contest_id}.}

\item{players}{A vector of player IDs to include. If \code{NULL} (the default), then use
all players found with \code{get_player_list()}.}

\item{exclude_players}{A vector of player IDs to exclude.}

\item{exclude_questionable}{Exclude players with statuses that indicate
they will not play. These include players that are questionable,
doubtful, out, and injured.}
}
\description{
Retrieve contest player list and other information used in
optimization.
}
\examples{
  \dontrun{
    get_contest_schematic(contest_id = 133645678)
  }

}
