% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{get_contests}
\alias{get_contests}
\title{Get List of Contests}
\usage{
get_contests(
  sport = NULL,
  throttle_rate = NULL,
  proxy_args = NULL,
  headers = NULL,
  retry_options = NULL,
  error_handling_options = NULL,
  curl_options = NULL,
  output = c("cleaned_json", "json", "response", "request")
)
}
\arguments{
\item{sport}{character. optional.}

\item{throttle_rate}{List of arguments to \link[httr2:req_throttle]{httr2::req_throttle}}

\item{proxy_args}{List of arguments to \link[httr2:req_proxy]{httr2::req_proxy}}

\item{headers}{List of arguments to \code{\link[httr2:req_headers]{httr2::req_headers()}}}

\item{retry_options}{List of arguments passed to \code{\link[httr2:req_retry]{httr2::req_retry()}}}

\item{error_handling_options}{List of arguments to \code{\link[httr2:req_error]{httr2::req_error()}}}

\item{curl_options}{List of arguments to \link[httr2:req_options]{httr2::req_options}}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}
}
\description{
Fetch the full table of contests and related info from DraftKings.com lobby
}
