% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftgroup.R
\name{get_draft_group}
\alias{get_draft_group}
\title{Get Draft Group Data}
\usage{
get_draft_group(draft_group_id = NULL, contest_id = NULL)
}
\arguments{
\item{draft_group_id}{Sequence of digits that correspond to a draft table/group.
If `draft_group_id` and `contest_id` are both passed, `contest_id` is ignored.}

\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}
}
\description{
Fetch information on the available lineup for a specific
  competition. In particular, this includes player names/IDs and
  their salaries.
}
