% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{get_gametype_rules}
\alias{get_gametype_rules}
\title{Get Gametype Rules}
\usage{
get_gametype_rules(game_type_id = NULL, contest_id = NULL)
}
\arguments{
\item{game_type_id}{Integer corresponding to the game type.
For example, 159 in \url{https://api.draftkings.com/lineups/v1/gametypes/159/rules}.
If both `game_type_id` and `contest_id` are passed, then `contest_id` is ignored.}
}
\description{
Fetch rules corresponding to a specific game type ID.
}
