% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_optimal_lineups}
\alias{get_optimal_lineups}
\title{Get Optimal Lineups}
\usage{
get_optimal_lineups(
  contest_schematic = NULL,
  contest_id = NULL,
  n = 1,
  exclude_players = NULL,
  tolerance = 0.001
)
}
\arguments{
\item{contest_schematic}{Output from \verb{[get_contest_schematic()]} which
includes information needed for optimization; including player info such as,
news status, salary, and projected fantasy points.}

\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{n}{Number of lineups to return}

\item{tolerance}{The difference in fantasy points between the previous
optimal lineup and the current optimal lineup during iterations.
Usually the default of 0.001 is fine.}
}
\description{
Wrapper around \code{optimize_lineup()}. Can return an
arbitrary number of optimal lineups in succession.
}
