% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sports.R
\name{get_sports}
\alias{get_sports}
\title{Get Sports}
\usage{
get_sports(output = c("cleaned_json", "json", "response", "request"), ...)
}
\arguments{
\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{Arguments passed to \code{\link[=dk_request]{dk_request()}}}
}
\description{
Fetch list of sport names and IDs.
}
\examples{
  \dontrun{
    get_sports()
  }

}
