% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_multi_get.R
\name{multi_req_perform_with_sleep}
\alias{multi_req_perform_with_sleep}
\title{Perform multiple HTTP requests with a delay}
\usage{
multi_req_perform_with_sleep(reqs, sleep = 3, ...)
}
\arguments{
\item{reqs}{A list of \link[httr2]{request}s.}

\item{sleep}{A numeric value indicating the number of seconds to sleep
after performing the requests. If NULL or negative, no sleep is performed.}

\item{...}{Other arguments passed to \code{\link[httr2:multi_req_perform]{httr2::multi_req_perform()}}}
}
\value{
A list of responses from \code{httr2::multi_req_perform}.
}
\description{
\code{multi_req_perform_with_sleep} performs multiple HTTP requests concurrently
and introduces a delay (sleep) after the requests have been sent.
}
\details{
This function is a wrapper around \code{httr2::multi_req_perform} that allows for
the introduction of a delay after sending out a batch of requests. This can be
useful for rate limiting or when dealing with APIs that have request quotas
over time.
}
\examples{
req_list <- list(
  httr2::request("http://httpbin.org/get"),
  httr2::request("http://httpbin.org/get")
)
responses <- multi_req_perform_with_sleep(req_list, sleep = 2)

}
\keyword{internal}
