% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimize_lineup}
\alias{optimize_lineup}
\title{Optimize Player Lineup}
\usage{
optimize_lineup(
  ...,
  contest_schematic = NULL,
  contest_id = NULL,
  max_points = NULL
)
}
\arguments{
\item{contest_schematic}{Output from `[get_contest_schematic()]` which
includes information needed for optimization; including player info such as,
news status, salary, and projected fantasy points.}

\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{max_points}{Optional upper threshold of points. This is used primarily
to get many successive lineups in `[get_optimal_lineups()]`}
}
\description{
Determine the optimal lineup of players based on the list of
available players, their salaries, and their projected
fantasy points.
}
