% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{pretty_duration}
\alias{pretty_duration}
\title{Format duration between two times in seconds, minutes, hours, or days}
\usage{
pretty_duration(start_time, end_time = Sys.time(), prefix = NULL)
}
\arguments{
\item{start_time}{The start time as a POSIXct object.}

\item{end_time}{The end time as a POSIXct object.}

\item{prefix}{Optional text to include before the duration.}
}
\value{
A character string representing the duration between the start and end
time, formatted as an integer followed by the appropriate time unit (e.g.
"1 day", "30 minutes", "5 seconds").
}
\description{
This function takes the start and end time as inputs and outputs a pretty
formatted duration in seconds, minutes, hours, or days.
}
\examples{

\dontrun{
start_time <- as.POSIXct("2023-03-17 12:00:00", tz = "UTC")
end_time <- as.POSIXct("2023-03-18 12:30:45", tz = "UTC")
pretty_duration(start_time, end_time)
}

}
\keyword{internal}
