% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{write_lineups_to_csv}
\alias{write_lineups_to_csv}
\title{Write Lineups to CSV}
\usage{
write_lineups_to_csv(lineups, file = "lineups.csv")
}
\arguments{
\item{lineups}{Output from \verb{[get_optimal_lineups()]}}

\item{file}{Path and name of the CSV file to create}
}
\description{
Given the output from \verb{[get_optimal_lineups()]}, create a
CSV formatted for uplaod to \url{https://www.draftkings.com/lineup/upload}
}
