% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-wrappers.R
\name{add_curl_options}
\alias{add_curl_options}
\title{Add Curl Options}
\usage{
add_curl_options(req, curl_options)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{curl_options}{List of arguments to \link[httr2:req_options]{httr2::req_options}}
}
\description{
Add list of curl arguments to a request object
}
