% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-wrappers.R
\name{add_proxy}
\alias{add_proxy}
\title{Add Proxy Args}
\usage{
add_proxy(req, proxy_args)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{proxy_args}{List of arguments to \link[httr2:req_proxy]{httr2::req_proxy}}
}
\description{
Add list of proxy arguments to a request object
}
