% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-wrappers.R
\name{add_retry}
\alias{add_retry}
\title{Add Retry}
\usage{
add_retry(req, retry_options)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{retry_options}{List of arguments passed to \code{\link[httr2:req_retry]{httr2::req_retry()}}}
}
\description{
Add list of arguments to a request object that control when a request is retried
}
