% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_throttle}
\alias{add_throttle}
\title{Add Throttle}
\usage{
add_throttle(req, throttle_rate)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{throttle_rate}{List of arguments to `httr2::req_throttle`}
}
\description{
Add list of arguments for `httr2::req_throttle` to a request object
}
