% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaderboard.R
\name{dk_get_entries}
\alias{dk_get_entries}
\title{Get Entry Details}
\usage{
dk_get_entries(
  draft_group_id,
  entry_keys,
  cookiefile = path.expand("~/cookies.txt"),
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{draft_group_id}{Sequence of digits that correspond to a draft table/group.
If \code{draft_group_id} and \code{contest_key} are both passed, \code{contest_key} is ignored.}

\item{entry_keys}{Vector of numeric (or character) keys that correspond to a specific entry in
a specific contest. See output from \code{\link[=dk_get_leaderboard]{dk_get_leaderboard()}}.}

\item{cookiefile}{Path to the cookies needed to perform API request.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{
  Arguments passed on to \code{\link[=dk_get_contest_info]{dk_get_contest_info}}
  \describe{
    \item{\code{contest_key}}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}
  }}
}
\description{
Fetch details for entries in a contest,
including the drafted roster for each
entry, stats for each player in an entry roster,
and the fantasy points associated to each stat.
}
\examples{
  \dontrun{
    dk_get_entries(draft_group_id = 80584, entry_keys = c(3618408508, 3618897002))
  }

}
