% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaderboard.R
\name{dk_get_leaderboard}
\alias{dk_get_leaderboard}
\title{Get Leaderboard}
\usage{
dk_get_leaderboard(
  contest_id,
  cookiefile = path.expand("~/cookies.txt"),
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{cookiefile}{Path to the cookies needed to perform API request.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{
  Arguments passed on to \code{\link[=dk_get_contest_info]{dk_get_contest_info}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Fetch leaderboard for a contest. A leaderboard
contains a list of entries in a contest,
fantasy points for each entry, associated user
for each entry, rank, and winnings.
}
\examples{
  \dontrun{
    dk_get_leaderboard(133645678)
  }

}
