% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_multi_get.R
\name{dk_multi_get}
\alias{dk_multi_get}
\title{Get Draft Kings Data in Parallel}
\usage{
dk_multi_get(
  reqs,
  paths = NULL,
  chunk_size = NULL,
  n_chunks = NULL,
  batch_sleep = 3,
  retry_sleep = 3,
  report_time = TRUE,
  .progress = TRUE,
  ...
)
}
\arguments{
\item{reqs}{A list of \link[httr2]{request}s.}

\item{paths}{An optional list of paths, if you want to download the request
bodies to disks. If supplied, must be the same length as \code{reqs}.}

\item{chunk_size, n_chunks}{Arguments passed to \code{\link[=chunk]{chunk()}}.
By default, \code{reqs} is not batched into chunks.}

\item{batch_sleep}{Delay in seconds between batches defined by \code{chunk_size} or \code{n_chunks}.}

\item{retry_sleep}{Delay in seconds between retries. See \code{\link[=multi_req_perform2]{multi_req_perform2()}} for details.}

\item{report_time}{Should the time it took the function to complete be reported?}

\item{.progress}{Whether to show a progress bar. Use \code{TRUE} to turn on
a basic progress bar, use a string to give it a name, or see
\link[purrr]{progress_bars} for more details.}

\item{...}{Other arguments passed to \code{\link[=multi_req_perform2]{multi_req_perform2()}}.}
}
\value{
A list of \code{httr2} response objects returned by \code{\link[httr2:multi_req_perform]{httr2::multi_req_perform()}}
}
\description{
Fetch DraftKings data in parallel using \code{httr2::multi_req_perform()}
with retry logic and in batches.
}
