% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{dk_optimize_lineup}
\alias{dk_optimize_lineup}
\title{Optimize Lineup}
\usage{
dk_optimize_lineup(contest_schematic, ...)
}
\arguments{
\item{contest_schematic}{Output from \verb{[get_contest_schematic()]} which
includes information needed for optimization; including player info such as,
news status, salary, and projected fantasy points.}

\item{...}{Other arguments passed to optimization method.}
}
\description{
Get optimal players based on player pool, projected fantasy points,
and rules defined by a contest schematic.
}
