% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{dk_write_csv}
\alias{dk_write_csv}
\alias{dk_write_csv.showdown_captain_mode_multiple_solutions}
\title{Write Lineups to CSV}
\usage{
dk_write_csv(optimal_lineups, file, ...)

\method{dk_write_csv}{showdown_captain_mode_multiple_solutions}(optimal_lineups, file = "showdown_captain_mode_lineups.csv", ...)
}
\arguments{
\item{optimal_lineups}{Output from \code{\link[=dk_get_optimal_lineups]{dk_get_optimal_lineups()}}}

\item{file}{Path and name of the CSV file to create}

\item{...}{Other args passed to \code{\link[utils:write.table]{utils::write.csv()}}}
}
\description{
Create CSV of optimal lineups that can be uploaded to
\url{https://www.draftkings.com/lineup/upload}.

Given the output from \code{\link[=dk_get_optimal_lineups]{dk_get_optimal_lineups()}}, create a
CSV of Showdown Captain Mode lineups formatted for
uplaod to \url{https://www.draftkings.com/lineup/upload}.
}
