% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkreq.R
\name{dkreq}
\alias{dkreq}
\title{Create Draft Kings Request Object}
\usage{
dkreq(
  proxy_args = NULL,
  curl_options = NULL,
  throttle_rate = NULL,
  headers = NULL,
  paths = NULL,
  query_params = NULL,
  retry_options = NULL,
  error_handling_options = NULL,
  base_url = "https://api.draftkings.com/"
)
}
\arguments{
\item{proxy_args}{List of arguments to \link[httr2:req_proxy]{httr2::req_proxy}}

\item{curl_options}{List of arguments to \link[httr2:req_options]{httr2::req_options}}

\item{throttle_rate}{List of arguments to \link[httr2:req_throttle]{httr2::req_throttle}}

\item{headers}{List of arguments to \code{\link[httr2:req_headers]{httr2::req_headers()}}}

\item{paths}{List of arguments to \code{\link[httr2:req_url]{httr2::req_url_path_append()}}}

\item{query_params}{A list of query parameters passed to \verb{[httr2::req_url_query]}.}

\item{retry_options}{List of arguments passed to \code{\link[httr2:req_retry]{httr2::req_retry()}}}

\item{error_handling_options}{List of arguments to \code{\link[httr2:req_error]{httr2::req_error()}}}

\item{base_url}{Base URL for request.}
}
\description{
Create Draft Kings Request Object
}
