% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dkreq_process}
\alias{dkreq_process}
\title{Process Request Based on Output Argument}
\usage{
dkreq_process(req, output)
}
\arguments{
\item{req}{A \link[httr2]{request}.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then `tidyjson::spread_all` is used
to parse the JSON body,}
}
\description{
Process Request Based on Output Argument
}
