% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{get_contest_info}
\alias{get_contest_info}
\title{Get Contest Info}
\usage{
get_contest_info(
  contest_id,
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{Arguments passed to \code{\link[=dk_request]{dk_request()}}}
}
\description{
Fetch contest information such as the sport, payout, and contest summary.
}
\examples{
  \dontrun{
    get_contest_info(133645678)
  }

}
