% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{get_contest_info}
\alias{get_contest_info}
\title{Get Contest Info}
\usage{
get_contest_info(
  contest_id,
  throttle_rate = NULL,
  proxy_args = NULL,
  headers = NULL,
  retry_options = NULL,
  is_error_func = NULL,
  curl_options = NULL,
  output = c("cleaned_json", "json", "response", "request")
)
}
\arguments{
\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{throttle_rate}{List of arguments to `httr2::req_throttle`}

\item{proxy_args}{List of arguments to `httr2::req_proxy`}

\item{headers}{List of arguments to `httr2::req_headers`}
}
\description{
Fetch contest information such as the sport, payout, and contest summary.
}
\examples{
  \dontrun{
    get_contest_info(133645678)
  }

}
