% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_contest_schematic}
\alias{get_contest_schematic}
\title{Get Contest Schematic}
\usage{
get_contest_schematic(
  contest_id,
  exclude_players = NULL,
  exclude_questionable = TRUE
)
}
\arguments{
\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{exclude_players}{Optional list of full player names to exclude}

\item{exclude_questionable}{Exclude players with statuses that indicate
they will not play. These include players that are questionable,
doubtful, out, and injured.}
}
\description{
Retrieve contest player list and other information used in
  optimization.
}
