% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftgroup.R
\name{get_draft_group_info}
\alias{get_draft_group_info}
\title{Get Draft Group Info}
\usage{
get_draft_group_info(
  draft_group_id = NULL,
  contest_id = NULL,
  output = c("cleaned_json", "json", "response", "request"),
  ...
)
}
\arguments{
\item{draft_group_id}{Sequence of digits that correspond to a draft table/group.
If \code{draft_group_id} and \code{contest_id} are both passed, \code{contest_id} is ignored.}

\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}

\item{output}{One of "cleaned_json" (the default),
"json", "response", or "request". If "cleaned_json" then \code{\link[tidyjson:spread_all]{tidyjson::spread_all()}} is used
to parse the JSON body,}

\item{...}{
  Arguments passed on to \code{\link[=get_draftable_players]{get_draftable_players}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Fetch info for a specific draft group
}
