% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftgroup.R
\name{get_player_list}
\alias{get_player_list}
\title{Get Player List}
\usage{
get_player_list(draft_group_id = NULL, contest_id = NULL)
}
\arguments{
\item{draft_group_id}{Sequence of digits that correspond to a draft table/group.
If \code{draft_group_id} and \code{contest_id} are both passed, \code{contest_id} is ignored.}

\item{contest_id}{The sequence of digits that correspond to a specific contest.
This can be found by examining the URL of a contest page.
For example: \url{https://www.draftkings.com/draft/contest/133645678#}. Here the contest ID
is 133645678.}
}
\description{
Fetch list of players for a specific draft group and related info
}
