% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftgroup.R
\name{get_player_points}
\alias{get_player_points}
\title{Get Player Points Earned}
\usage{
get_player_points(year, week, sport = "nfl")
}
\arguments{
\item{year}{integer.}

\item{week}{integer.}

\item{sport}{character.}
}
\description{
Retrieve player fantasy points earned in each
game for a given season and week.
}
